Add config.py
This module will hold path constants to typical files.
It also handles adjusting the path once I freeze this program as an EXE which requires leveraging the MEIPASS value in Windows.
Lastly, `_fix_paths` is to ensure the logging the handler has a fully qualified file path to LOGS_DIR for its given filename.