Add Settings parser
This will parse application settings from settings.yaml
The most important setting will be the install_path.
The default install path is provided/assumed.
However, there are some helper methods to ensure the value provided is a proper path, exists, and is a Path object for simplicity later.
The parser will be loaded so Settings can be imported to other modules fully initialized.