Create reusable workflows (#9)
This is largely still WIP, but will serve as a jumping off point since I can't test the workflows themselves without getting them into `main` to trigger off of initially first. I likely could have just added them as blank files, but that was an after thought and I was still fleshing out the implementation. So, here we are.